refactor: expose parser client interface at the parser base
